Workaround nv_vm_flags_* calling GPL-only code
Since 6.15, the main body of vma_start_write() was uninlined and has
been made GPL-only code. This means that the closed modules can no
longer use vma_start_write() and functions calling it. Fortunately for
us, an NVIDIA driver hacker that tries to maintain the legacy drivers
for newer kernels claims that all VMAs that the driver uses are all
either initially mapped or already locked, this means that using
vm_flags_reset is sufficient to replace vm_flags_* helpers.

Signed-off-by: Eric Naim <dnaim@cachyos.org>